move base.rs into api.rs
No real change here, just consolidating files because I keep clicking on
the wrong one. :-)
