+++ src/select/sel_code_block.rs
@@ -1,7 +1,6 @@
 use crate::matcher::{CharEnd, StringMatcher};
 use crate::parsing_iter::ParsingIterator;
-use crate::select::base::Selector;
-use crate::select::{ParseResult, SELECTOR_SEPARATOR};
+use crate::select::{ParseResult, Selector, SELECTOR_SEPARATOR};
 use crate::tree::{CodeBlock, CodeVariant};
 
 #[derive(Debug, PartialEq)]
